chore: move db package into core
