+++ src/cmipld/core/service/settings_default.toml
@@ -7,7 +7,7 @@
 # db_path = "./cache/dbs/XXXX.sqlite"
 
 [[projects]]
-project_name = "CMIP6Plus_CVs"
+project_name = "cmip6plus"
 github_repo = "https://github.com/ESPRI-Mod/CMIP6Plus_CVs"
 branch = "uni_proj_ld"
 local_path = ".cache/repos/CMIP6Plus_CVs"
